Refactor render function
Moved out a lot of static code from the render pipeline.
These weren't complex calculations, but already we can see how
massively the CPU can bottleneck the GPU. With this change the
frame times went from about 700 microseconds to 200 microseconds.
Obviously this is already from a 1.4k framerate to a 5k framerate,
which is ridiculous, but that isn't the point.
The lesson seems to be, that in the render pass, you don't want to
allocate memory, do any CPU intensive operations. The data should be
right there, ready to go, to be blasted to the hungry hungry GPU.