Extend projection support to 3D
Left in how you would construct a projection (to_clip_space) matrix
by hand and implemented a function which returns the same thing.
We arbitrarily choose that the world space is 400 pixels deep, but
the clip space will only go between -1.0, 1.0. Instead of the -2.0, 2.0
as for the other axes.

The other thing of note is that vertex buffers are weird. Notice that
for each vertex we are allowed to supply only 3 floats and the GPU
will automatically assume that the 4 float (w) is of value 1.0, hence
transforming a 3d vector to a 4d vector.
This is something to keep in mind, as we can save 1/4th of the
bus bandwitdth.