Extend mul trait with reference based variants
The goal is to support all these operations:
Vector<T> * T
Vector<T> * &T
&Vector<T> * T
&Vector<T> * &T
The rest of the possible compinations produced by variants of T
itself should be handled by the regular template substitution.

Why the last 3 options?
A library should impose the least amount of limitations and it should
not be necessary to force the user on cloniing/copying their T when
they want to execute an operation.
There may be a cleaner way to do this, but for now, as a start it
suffices.