Refactor std::ops::Mul/MulAssign for lina::Vector
After finally managign to use uninitialized memory for the std::ops::Div
implementation it was clear that there was some misunderstanding
regarding Copy and referencedness.
For some reason I believed that
Vector<T> * &T
&Vector<T> * &T
etc implementations would be helpful. But in reality they serve only
to obfuscate the underlying operations.
&Vector<T> * &T looks like we are taking the components by reference
and not copying any of them, but this isn't the case in reality.
When we get down to the implementation itself we realize that
at some point std::ops::Mul<T> for T has to be implemented.
Which takes by value, so it will destroy the input at the bottom level
either way.

Furthermore since uninitialized memory requires Copy to be implemented
we might as well require it for the lina::Vector class itself. With
this it was really clear that the operators were implemented
incorrectly.