Impl cross product with macro and move all macros into macro.rs
There is a peculiar issue that I wasn't aware off before.
It goes like this:
Cross product macro wants to use the std::ops::mul generated by the
lhs_scalar_mul_impl macro. The rust compiler will run some bounds
checking which makes it enter into a circular loop, if and only
if the two macro implementations are in a separate module.
The problem is completely non-existent if all macro calls are in
the same file. But to not to have problems with exporting the
macro defitions, those were moved into a single file as well.