Start using the vector/matrix library
A few indexing traits were missing, those were added as well.
Now all that temporary code could be removed. Ahh, glorious!
The only thing that was left behind is the inverse matrix calculation.
Don't like it, but a general implementation may be problematic, so for
the time being it is left there.