Clean up point_at and look_at matrices
Unfortunatelly I still find it hard to understand what is exactly
happening, even though a considerable time has been spent on the
problem.
This is why I didn't document my findings any further.
Understand the general principles, but too many tricks are compressed
too tightly for me to be able to properly explain it.
For now it works, when the problem arrises we will look into it again.