Add a quick and dirty constant rotation to the cube
For a long while this little cube will be sufficient to
see if we are on the right path when implementing some features.
Until then it doesn't matter if it is a little dirty.
Of course ideally we would have scenes and scene graphs and all that
jazz, but as a start it isn't necessary.
(More importantly I am unsure how they would even look like anyways)