Add perspective projection matrices
One general that handles asymmetric frustra, one for symmetric only
and variants for both with infinite z_far.
It was rather not trivial to make these.
Documentation will be added later as there is still one or two common
variant that need to be implemented.
Some of them may not have been necessary, but it made more sense to
implement them now, and not touch them later.