Turn Vector<T, 3>::cross into a partially specified template
Originally it was a macro, as the cross product can only
be defined in 3 and 7 dimensions anyways, which confused me
into believing that a macro was the proper choice.

With a templated implementation other libs can also relly on the
existence of the cross product, but this isn't true with a macro
implementation.