Implement a simple spot light
My oh my, memory padding is problematic and very error prone.
Will have to deal with that at some point.

We know see that a spot light or an even simpler directional light
is really nothing just a vector and some values on the CPU side,
and some minimal shader code on the GPU side.
Stating this, because I am thinking about some more generalized patterns
on how these can be implemented and uploaded to the GPU.