Add support for multiple key presses
Now multiple keyboard presses are properly registered by the application
and the speed of change is scalled according to the framerate and
the defined base speed.
This provides a nice and smooth navigation experience in contrast
to the previous jerky one (didn't even notice it really, was
so focused on the rendering aspects).

Also added support for the Shift key, which in will tripple the move
speed when held. Not sure how much would I use it, but Godot
does something similar, so it is probabbly a good idea.